+++ Modules/repetition.js
@@ -2,10 +2,10 @@ import { Book } from "./book.js";
 
 const msg = document.querySelector("#errorMsg")
 const deleteBtn = document.createElement("button")
-const inpAuthor = document.querySelector("#author");
-const inpTitle = document.querySelector("#title");
+// const inpAuthor = document.querySelector("#author");
+// const inpTitle = document.querySelector("#title");
 
-const checkRepetition = book => {
+export const checkRepetition = book => {
   for (let i = 0; i < Book.collection.length; i += 1) {
     const currentBook = Book.collection[i]
     if (
@@ -27,12 +27,12 @@ deleteBtn.addEventListener("click", () => {
   msg.classList.add("hidden")
 })
 
-export const addBookToLibrary = () => {
-  const inputTitleValue = inpTitle.value
-  const inputAuthorValue = inpAuthor.value
-  const book = new Book(inputTitleValue, inputAuthorValue)
-  if (checkRepetition(book)) {
-    book.saveBooks()
-  }
-}
+// export const addBookToLibrary = () => {
+//   const inputTitleValue = inpTitle.value
+//   const inputAuthorValue = inpAuthor.value
+//   const book = new Book(inputTitleValue, inputAuthorValue)
+//   if (checkRepetition(book)) {
+//     book.saveBooks()
+//   }
+// }
